+++ addon/components/cedar-chart/component.js
@@ -1,10 +1,12 @@
 /* globals Cedar:false */
 import Ember from 'ember';
-import layout from './template';
+
+// test if an object is empty ({})
+function isEmptyObject (obj) {
+  return obj && obj.constructor === Object && Object.keys(obj).length === 0;
+}
 
 export default Ember.Component.extend({
-  layout: layout,
-
   // show chart at root DOM elememt of this component
   _showChart() {
     if (this.isDestroyed || this.isDestroying) {
@@ -16,28 +18,9 @@ export default Ember.Component.extend({
 
     // create and show the chart
     try {
-      const spec = this.get('specification');
-      if (!spec) {
-        return;
-      }
-
-      // Copy spec so that we don't mutate this.get('specification')
-      // Re-evaluate once we resolve https://github.com/Esri/cedar/issues/211 in cedar
-      const specification = Ember.$.extend(true, {}, spec);
-      const options = this.get('options') || {};
-      const override = this.get('override') || options.override;
-
-      // use elementId from component to render Cedar directly into div
-      options.elementId = '#' + this.elementId;
-
-      // autolabels will throw an error on pie charts
-      // see https://github.com/Esri/cedar/issues/173
-      if (specification.type === 'pie') {
-        options.autolabels = false;
-      }
-
-      // create the chart
-      this.chart = new Cedar(specification);
+      // get chart constructor options from properties and create the chart
+      const props = this.getProperties('type', 'dataset', 'specification', 'tooltip', 'override', 'transform', 'timeout');
+      this.chart = new Cedar(props);
 
       // wire up event handlers
       const supportedEvents = ['Click', 'Mouseover', 'Mouseout', 'Mousemove', 'UpdateStart', 'UpdateEnd'];
@@ -49,8 +32,15 @@ export default Ember.Component.extend({
         }
       });
 
-      // wire up data transform function if any
-      this.chart.transform = this.transform;
+      // get render options from properties but use elementId from component
+      const options = this.get('options') || {};
+      options.elementId = '#' + this.elementId;
+
+      // autolabels will throw an error on pie charts
+      // see https://github.com/Esri/cedar/issues/173
+      if (this.chart.type === 'pie') {
+        options.autolabels = false;
+      }
 
       // attach the chart to the DOM
       this.chart.show(options, err => {
@@ -59,11 +49,6 @@ export default Ember.Component.extend({
           this._handleErr(err);
         }
       });
-
-      // look for overrides & apply
-      if (override) {
-        this.chart.override = override;
-      }
     }
     catch(err) {
       // an error occurred while creating the cart
@@ -91,12 +76,49 @@ export default Ember.Component.extend({
     }
   },
 
-  didReceiveAttrs(e) {
-    // now we call an error handler instead of showing a static message
-    // when there are chart errors 
-    if (e.newAttrs.invalidSpecMessage) {
-      console.warn('DEPRECATION: Usage of `invalidSpecMessage` is deprecated, pass an action to `onError` instead.');
+  _deprecate (oldKey, newPropertyName, oldContext) {
+    const oldPropertyName = oldContext ? oldContext + '.' + oldKey : oldKey;
+    const oldPropertyValue = this.get(oldPropertyName);
+    if (oldPropertyValue) {
+      this.set(newPropertyName, oldPropertyValue);
+      if (oldContext) {
+        delete this.get(oldContext)[oldKey];
+      } else {
+        delete this[oldPropertyName];
+      }
+      this._showDeprecationWarning(oldPropertyName, newPropertyName);
     }
+  },
+
+  // emulate Ember deprecation console warnings
+  // for nested properties, etc that we can't use deprecatingAlias for
+  _showDeprecationWarning (oldPropertyName, newPropertyName) {
+    console.warn('DEPRECATION: Usage of `' + oldPropertyName + '` is deprecated, use `' + newPropertyName + '` instead.');
+  },
+
+  didReceiveAttrs (e) {
+    this._super(...arguments);
+    // now we call an error handler instead of showing a static message
+    // when there are chart errors
+    if (e.newAttrs.invalidSpecMessage) {
+      this._showDeprecationWarning('invalidSpecMessage', 'onError');
+    }
+
+    // in ember-cli-cedar <= v0.5.0 Cedar constructor options
+    // were passed in via the specification property
+    // need to pull those out into own properties and show deprecation warning
+    this._deprecate('type', 'type', 'specification');
+    this._deprecate('dataset', 'dataset', 'specification');
+    this._deprecate('tooltip', 'tooltip', 'specification');
+    this._deprecate('timeout', 'timeout', 'specification');
+    this._deprecate('override', 'override', 'specification');
+    if (isEmptyObject(this.get('specification'))) {
+      delete this.specification;
+    }
+
+    // at one point in the distant past we sent in override via options
+    this._deprecate('override', 'override', 'options');
+
     // re-create and show chart whenever attributes change
     Ember.run.scheduleOnce('afterRender', this, '_showChart');
   },