+++ tests/dummy/app/components/charts-playground/component.js
@@ -4,15 +4,12 @@ import layout from './template';
 export default Ember.Component.extend({
   layout: layout,
 
-  tagName: 'div',
-  classNames: [],
-
-  specificationString: Ember.computed('model', 'model.specification', function() {
-    return JSON.stringify(this.get('model.specification'));
+  jsonString: Ember.computed('model', function() {
+    return JSON.stringify(this.get('model'), undefined, 2);
   }),
   actions: {
     renderJSON: function() {
-      this.set('model.specification', JSON.parse(this.get('specificationString')));
+      this.set('model', JSON.parse(this.get('jsonString')));
     }
   }
 });