+++ tests/dummy/app/transform/route.js
@@ -4,6 +4,11 @@ export default Ember.Route.extend({
   model () {
     return {
       type: 'pie',
+      options: {
+        width: 600,
+        height: 600,
+        autolabels: false
+      },
       dataset: {
         "url":"https://services.arcgis.com/uDTUpUPbk8X8mXwl/arcgis/rest/services/Public_Schools_in_Onondaga_County/FeatureServer/0",
         "query": {